performance comarison with js function

+++ src/factorial.h
@@ -4,7 +4,7 @@ using namespace Napi;
 Napi::Value Factorial(const Napi::CallbackInfo& info) {
   Napi::Env env = info.Env();
   int arg0 = info[0].As<Napi::Number>().DoubleValue();
-  long int res = 1, i; 
+  long long int res = 1, i; 
   for (i = 2; i <= arg0; i++) res *= i;
   return Napi::Number::New(env, res);
 }
